% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_check.R
\name{density_check}
\alias{density_check}
\title{Internal function for density check of check layout}
\usage{
density_check(layout.matrix, by.density)
}
\arguments{
\item{layout.matrix}{The layout matrix of designed grid}

\item{by.density}{The check density in the layout matrix}
}
\description{
This is an inernal function
}
