% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_pattern.R
\name{diag_extract}
\alias{diag_extract}
\title{Extract the designed diagonal pattern from the system data file (Internal functions)}
\usage{
diag_extract()
}
\description{
Functions of extracting the pre-designed diagonal pattern (density from 0.01 up to 0.35)
}
