% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal_check_by_density.R
\name{diagonal_check_by_density}
\alias{diagonal_check_by_density}
\title{Generate a spatial layout of checks by diagonal design}
\usage{
diagonal_check_by_density(n.row, n.col, by.density, spread.scale = 50 *
  max(n.row, n.col))
}
\arguments{
\item{n.row}{The row of the retangular grid}

\item{n.col}{The column of teh retangular grid}

\item{by.density}{The density of the check}

\item{spread.scale}{The spreading of the check across diagonal, related to the scale parameter in Cauchy distribution, default=50*max(n.row, n.col)}
}
\description{
This funciton generates a layout grid with checks on diagonal
}
\examples{
#Creat a 30 by 30 layout with 0.2 density
tmp = diagonal_check_by_density(30, 30, 0.2)
fields::quilt.plot(tmp, nx=30, ny=30)
}
