% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagSet_check_by_density.R
\name{diagSet_check_by_density}
\alias{diagSet_check_by_density}
\title{Generate a spatial layout of checks by diagonal design (settle for density = .1 - .35}
\usage{
diagSet_check_by_density(n.row, n.col, by.density)
}
\arguments{
\item{n.row}{The row of the retangular grid}

\item{n.col}{The column of the retangular grid}

\item{by.density}{The density of the checks, range from 0.01 to 0.35}
}
\description{
This function generates a layout grid with checks on diagonal (pre-designed)
}
\examples{
nr=10; nc=20; den=0.2
diagSet_check_by_density(n.row=nr, n.col=nc, by.density=den)
}
