% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_check_by_plot.R
\name{equal_check_by_plot}
\alias{equal_check_by_plot}
\title{Generate a spatial layout of checks by every # of plots}
\usage{
equal_check_by_plot(n.row, n.col, by.plot)
}
\arguments{
\item{n.row}{The row of a retangular grid}

\item{n.col}{The column of a retangular grid}

\item{by.plot}{The distance between checks (row-wise and col-wise) in the unit of plots}
}
\value{
A dataframe will return
}
\description{
This function generate a check layout in a retangular grid, The check is distributed by every # of plots
}
\examples{
# To produce a 4 by 5 field with check distributed by every 2 plots
nr = 4; nc = 5
example.tmp = equal_check_by_plot(nr, nc, 2)
fields::quilt.plot(example.tmp, nx=nr, ny=nc)
}
