% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_space_check_by_density.R
\name{equal_space_check_by_density}
\alias{equal_space_check_by_density}
\title{Generate a spatial layout of checks by density}
\usage{
equal_space_check_by_density(n.row, n.col, by.density)
}
\arguments{
\item{n.row}{The row of a retangular grid}

\item{n.col}{The column of a retangular grid}

\item{by.density}{The density of the check}
}
\value{
A dataframe will return
}
\description{
This function generate a check layout in a retangular grid, The check is evenly distributed by the given density
}
\examples{
# To produce a 4 by 5 field with check distributed by density=0.2
nr = 4; nc = 5
example.tmp = equal_space_check_by_density(nr, nc, .2)
fields::quilt.plot(example.tmp, nx=nr, ny=nc)
}
