% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_spatial_gau.R
\name{iso_spatial_gau}
\alias{iso_spatial_gau}
\title{Isotropic spatial simulation by Gaussian function}
\usage{
iso_spatial_gau(dist.data, nugget, sill, ranges)
}
\arguments{
\item{dist.data}{the a distance matrix generate from spatial_dist function}

\item{nugget}{the nugget value of the var-cov function}

\item{sill}{the sill(variance, sigma^2) of the var-cov function}

\item{ranges}{the range value of var-cov function}
}
\value{
A matrix of var-cov matrix
}
\description{
This function load grid floor and generate the var-cov matrix via Gaussian function
}
\examples{
D = spatial_dist(expand.grid(1:5, 1:5))
iso_spatial_gau(D, 0, 3, 1)
}
