% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_check_by_density.R
\name{random_check_by_density}
\alias{random_check_by_density}
\title{Generate a random spatial layout of check}
\usage{
random_check_by_density(n.row, n.col, by.density)
}
\arguments{
\item{n.row}{The row of the retangular grid}

\item{n.col}{The column of the retangular grid}

\item{by.density}{The density of checks}
}
\description{
This function generate a random spatial layout of checks
}
\examples{
nr=nc=5
tmp = random_check_by_density(nr, nc, .1)
fields::quilt.plot(tmp, nx=nr, ny=nc)
}
