% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_dist.R
\name{spatial_dist}
\alias{spatial_dist}
\title{Distance matrix calculation}
\usage{
spatial_dist(M, lon.lat = F, mile = F)
}
\arguments{
\item{M}{A data.frame contains grid coordinates}

\item{lon.lat}{If TRUE, the function will use longitude and latitude for calculating geographic distance}

\item{mile}{If TRUE, the function will yeild distance in mileages. If FALSE, the function will yeild distance in kilometers}
}
\value{
A matrix of distance matrix
}
\description{
This function load grid floor and generate the distance matrix
}
\examples{
spatial_dist(expand.grid(1:4, 1:4))
}
